Fix signedness warning in gio/tests/contenttype.c
gio/tests/contenttype.c: In function ‘test_tree’:
gio/tests/contenttype.c:337:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’
  337 |   for (i = 0; i < G_N_ELEMENTS (tests); i++)
      |                 ^

+++ gio/tests/contenttype.c
@@ -363,7 +363,7 @@ test_tree (void)
   const gchar *path;
   GFile *file;
   gchar **types;
-  gint i;
+  gsize i;
 
 #ifdef __APPLE__
   g_test_skip ("The OSX backend does not implement g_content_type_guess_for_tree()");
